# docs/swagger.yaml
openapi : "3.0.0"
info:
  version: 3.5.0
  title: Datasance PoT Controller
paths:
...
  /agent/provision:
    post:
      tags:
        - Agent
      summary: Provision agent with an ioFog node
      operationId: agentProvision
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProvisioningRequest"
        required: true
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProvisioningResponse"
        "400":
          description: Bad Request
        "401":
          description: Expired Provisioning Key
        "404":
          description: Invalid Provisioning Key
        "500":
          description: Internal Server Error
  /agent/deprovision:
    post:
      tags:
        - Agent
      summary: Deprovision agent
      operationId: agentDeprovision
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDeprovisioningRequest"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/cert:
    get:
      tags:
        - Agent
      summary: Move Controller CA to Agent
      operationId: agentControllerCert
      security:
        - agentToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDeprovisioningRequest"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/config:
    get:
      tags:
        - Agent
      summary: Get an ioFog node configuration
      operationId: getIOFogNodeConfig
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOFogNodeConfig"
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
    patch:
      tags:
        - Agent
      summary: Updates an ioFog node configuration
      operationId: updateIOFogNodeConfig
      security:
        - agentToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IOFogNodeConfigRequest"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/config/changes:
    get:
      tags:
        - Agent
      summary: Gets ioFog node changes
      operationId: getIOFogNodeChanges
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOFogNodeConfigChanges"
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
    patch:
      tags:
        - Agent
      summary: Resets ioFog node changes list
      operationId: resetIOFogNodeChanges
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/status:
    put:
      tags:
        - Agent
      summary: Posts agent status to ioFog node
      operationId: postAgentStatus
      security:
        - agentToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStatus"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/microservices:
    get:
      tags:
        - Agent
      summary: Gets microservices running on an ioFog node
      operationId: getAgentMicroservicesList
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMicroservicesListResponse"
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  "/agent/microservices/{microserviceUuid}":
    get:
      tags:
        - Agent
      summary: Gets microservices running on an ioFog node
      operationId: getAgentMicroserviceInfo
      parameters:
        - in: path
          required: true
          name: microserviceUuid
          description: Microservice UUID
          schema:
            type: string
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMicroservicesInfoResponse"
        "401":
          description: Not Authorized
        "404":
          description: Invalid Microservice Uuid
        "500":
          description: Internal Server Error
  /agent/registries:
    get:
      tags:
        - Agent
      summary: Gets list of Docker registries
      operationId: getRegistriesList
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistriesListResponse"
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/tunnel:
    get:
      tags:
        - Agent
      summary: Get an ioFog node tunnel configuration
      operationId: getIOFogNodeTunnelConfig
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOFogNodeTunnelConfigResponse"
        "401":
          description: Not Authorized
        "404":
          description: Tunnel Not Found
        "500":
          description: Internal Server Error
  /agent/strace:
    get:
      tags:
        - Agent
      summary: Get an ioFog node strace info
      operationId: getIOFogNodeStraceInfo
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOFogNodeStraceResponse"
        "401":
          description: Not Authorized
        "404":
          description: Strace Not Found
        "500":
          description: Internal Server Error
    put:
      tags:
        - Agent
      summary: Posts agent strace to ioFog node
      operationId: postIOFogNodeStraceBuffer
      security:
        - agentToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IOFogNodeStraceBuffer"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "404":
          description: Invalid Node Id
        "500":
          description: Internal Server Error
...
  /agent/hal/hw:
    put:
      tags:
        - Agent
      summary: Updates HAL hardware info
      operationId: putHalHardwareInfo
      security:
        - agentToken: []
      requestBody:
        $ref: "#/components/requestBodies/HalInfo"
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/hal/usb:
    get:
      tags:
        - ioFog
      summary: Retrieves HAL USB info
      operationId: getAgentHalUsbInfo
      security:
        - userToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HalInfo"
        "401":
          description: Not Authorized
        "404":
          description: Invalid Node Id
        "500":
          description: Internal Server Error
    put:
      tags:
        - Agent
      summary: Updates HAL USB info
      operationId: putHalUsbInfo
      security:
        - agentToken: []
      requestBody:
        $ref: "#/components/requestBodies/HalInfo"
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/delete-node:
    delete:
      tags:
        - Agent
      summary: Deletes an ioFog node
      operationId: deleteAgentNode
      security:
        - agentToken: []
      responses:
        "204":
          description: No Content
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
  /agent/image-snapshot:
    get:
      tags:
        - Agent
      summary: Get image snapshot info
      operationId: getImageSnapshot
      security:
        - agentToken: []
      responses:
        "200":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSnapshotResponse"
        "401":
          description: Not Authorized
        "404":
          description: Image Snapshot Not Found
        "500":
          description: Internal Server Error
    put:
      tags:
        - Agent
      summary: Put image snapshot info on controller
      operationId: putImageSnapshot
      security:
        - agentToken: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageSnapshotRequest"
        required: true
      responses:
        "204":
          description: Success
          headers:
            X-Timestamp:
              description: FogController server timestamp
              schema:
                type: number
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
        "500":
          description: Internal Server Error
...
servers:
  - url: http://localhost:51121/api/v3
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AgentProvisioningRequest:
      type: object
      required:
        - type
        - key
      properties:
        type:
          type: number
          enum:
            - 1
            - 2
          description: |
            Architecture
             * '1': x86
             * '2': arm
        key:
          type: string
          description: provisioning key
    AgentProvisioningResponse:
      type: object
      properties:
        id:
          type: string
          description: ioFog Node Id
        token:
          type: string
          description: Agent token
    AgentDeprovisioningRequest:
      type: object
      properties:
        microserviceUuids:
          type: array
          items:
            type: string
    IOFogNodeConfigChanges:
      type: object
      properties:
        config:
          type: boolean
        version:
          type: boolean
        reboot:
          type: boolean
        deleteNode:
          type: boolean
        microservicesList:
          type: boolean
        microservicesConfig:
          type: boolean
        routing:
          type: boolean
        registries:
          type: boolean
        tunnel:
          type: boolean
        diagnostics:
          type: boolean
        isImageSnapshot:
          type: boolean
        prune:
          type: boolean
        routerChanged:
          type: boolean
    IOFogNodeConfig:
      type: object
      properties:
        networkInterface:
          type: string
        dockerUrl:
          type: string
        diskLimit:
          type: number
        diskDirectory:
          type: string
        memoryLimit:
          type: number
        cpuLimit:
          type: number
        logLimit:
          type: number
        logDirectory:
          type: string
        logFileCount:
          type: number
        statusFrequency:
          type: number
        changeFrequency:
          type: number
        deviceScanFrequency:
          type: number
        watchdogEnabled:
          type: boolean
        latitude:
          type: number
        longitude:
          type: number
        logLevel:
          type: string
        availableDiskThreshold:
          type: number
        dockerPruningFrequency:
          type: number
        routerHost:
          type: string
        routerPort:
          type: number
    IOFogNodeConfigRequest:
      type: object
      properties:
        networkInterface:
          type: string
        dockerUrl:
          type: string
        diskLimit:
          type: number
        diskDirectory:
          type: string
        memoryLimit:
          type: number
        cpuLimit:
          type: number
        logLimit:
          type: number
        logDirectory:
          type: string
        logFileCount:
          type: number
        statusFrequency:
          type: number
        changeFrequency:
          type: number
        deviceScanFrequency:
          type: number
        watchdogEnabled:
          type: boolean
        latitude:
          type: number
        longitude:
          type: number
        gpsMode:
          type: string
        logLevel:
          type: string
        availableDiskThreshold:
          type: number
        dockerPruningFrequency:
          type: number
...
    IOFogNodeStraceResponse:
      type: object
      properties:
        straceValues:
          type: array
          items:
            $ref: "#/components/schemas/MicroserviceStrace"
...
    AgentMicroservicesListResponse:
      type: object
      properties:
        microservices:
          type: array
          items:
            $ref: "#/components/schemas/AgentMicroservicesInfoResponse"
...
    RegistriesListResponse:
      type: object
      properties:
        registries:
          type: array
          items:
            $ref: "#/components/schemas/RegistryResponse"
...